backend/user/requirements.txt
mysqlclient==2.0.1
Django==3.1.2  
djangorestframework==3.12.1 
setuptools==44.0.0 
Unipath==1.1
pytest-django==3.9.0
pytest==5.4.1
django-mock-queries==2.1.5
pytest-mock==3.3.1
pytest-cov==2.10.1
pylint==2.6.0
requests==2.24.0
gunicorn==20.0.4